Sample Generation Intervals

Draws n samples from a lognormal distribution representing generation intervals (time between successive infections).

Usage

gi.dist.samp(n, meanlog = 1.376, sdlog = 0.567)

Arguments

n

Integer. Number of samples to draw.

meanlog

Numeric. Mean of the distribution on the log scale. Default is 1.376 (COVID-19 estimate, ~3.96 day median).

sdlog

Numeric. Standard deviation of the distribution on the log scale. Default is 0.567 (COVID-19 estimate).

Value

Numeric vector of length n containing generation intervals in days.

Details

The default parameters are calibrated to COVID-19 data. For other diseases, supply appropriate lognormal parameters:

Influenza: meanlog ≈ 0.91, sdlog ≈ 0.52 (~2.5 day median)

SARS: meanlog ≈ 2.0, sdlog ≈ 0.45 (~7.4 day median)

Measles: meanlog ≈ 2.4, sdlog ≈ 0.3 (~11 day median)

Examples

# COVID-19 default
gi.dist.samp(100)
#>   [1]  3.3726744  2.8923012  5.6555243 12.7670689  1.5702408  5.2938516
#>   [7]  1.3766768  2.9447352  3.8426980  5.3864090  2.3577778  5.1626167
#>  [13]  4.8636703  1.8895219  6.0153779 11.5511188  3.7462248  2.3288499
#>  [19]  3.9233904  2.4774026  1.6794551  6.7284912  4.3757111  4.5456451
#>  [25]  9.9398529  4.2186940  3.6693836  1.3404165  3.3793132  3.3143984
#>  [31]  7.2511753  4.1194092  2.7555500  3.8484478  3.4329120  5.0946958
#>  [37] 18.8841277  4.0648762  5.4934751  4.2334467  1.3391760  6.4546644
#>  [43]  3.4490015  3.5224211  4.0023180  4.0259502  5.4073126  1.0904352
#>  [49] 18.1198844  3.2258212  4.4681422  7.2801693  2.7152797  7.4455089
#>  [55]  3.4438042  2.0305803  2.2766219  7.2419282  4.2659174  5.2228986
#>  [61]  1.5104559  1.7196014  4.6511442  8.4507963  4.5276670  8.3601164
#>  [67]  5.3284311  5.5846745  3.7197872  4.3650386  3.7613744 11.7882429
#>  [73]  8.2663139  6.0530644  5.4269604  2.9012397  7.4310955  0.9001503
#>  [79]  3.6245191  5.0632851  3.1881274  5.0355077  7.2339038  7.1751241
#>  [85]  3.8744188  5.2155599 10.2218178  3.2383930  6.7705292  8.3531653
#>  [91]  3.3461322  3.1786288  2.5361817  2.1745684  2.5216867  1.4625651
#>  [97]  2.6763797  2.8843705  2.9203746  4.5031677

# Influenza (shorter generation interval)
gi.dist.samp(100, meanlog = 0.91, sdlog = 0.52)
#>   [1] 4.1321447 2.2286166 1.1999820 2.8418049 1.9743995 3.3390019 7.5079483
#>   [8] 3.0985256 1.0347642 2.8283371 4.3400067 7.1740620 3.1384089 5.1230159
#>  [15] 3.1012790 2.6272658 2.5132915 3.4003478 2.1671690 1.8875934 2.7453751
#>  [22] 1.3688571 3.8574873 2.5921670 1.2602866 1.5199018 3.1463959 1.5924834
#>  [29] 2.1400185 3.9565801 2.5728090 2.2828127 1.6157491 6.5915131 3.7007996
#>  [36] 4.1353728 4.9398063 1.3878425 3.2465468 1.1334523 5.5125404 3.1054438
#>  [43] 2.6471770 1.3746973 1.8586160 4.2944352 3.5338767 2.5345593 2.0640742
#>  [50] 3.7325025 3.7745986 0.9249352 4.0414673 2.1155043 2.8485061 0.9791134
#>  [57] 1.6488951 1.3782695 3.0014615 2.1411398 3.2517326 2.3548746 1.4970310
#>  [64] 4.8102526 4.0945208 3.7051921 4.2575080 1.9417288 1.2799562 2.1192818
#>  [71] 7.8469124 1.4453224 1.3686264 1.0395935 5.4930744 3.3140681 7.0035512
#>  [78] 2.2929913 9.4261933 4.3156138 4.5005578 4.4566448 2.0209329 1.6191547
#>  [85] 1.8385551 6.2162727 2.6622082 3.0215793 4.4912362 4.7372500 3.4153750
#>  [92] 1.9871915 5.0401704 2.3944514 2.1564861 0.6960919 2.5703782 1.4032281
#>  [99] 1.7873737 0.8495060