Sample Incubation Periods

Draws n samples from a lognormal distribution representing incubation periods.

Usage

inc.period.samp(n, meanlog = 1.63, sdlog = 0.5)

Arguments

n

Integer. Number of samples to draw.

meanlog

Numeric. Mean of the distribution on the log scale. Default is 1.63 (COVID-19 estimate, ~5.1 day median).

sdlog

Numeric. Standard deviation of the distribution on the log scale. Default is 0.5 (COVID-19 estimate).

Value

Numeric vector of length n containing incubation periods in days.

Details

The default parameters are fitted to COVID-19 data. For other diseases, supply appropriate lognormal parameters:

Influenza: meanlog ≈ 0.34, sdlog ≈ 0.42 (~1.4 day median)

SARS: meanlog ≈ 1.39, sdlog ≈ 0.51 (~4 day median)

MERS: meanlog ≈ 1.61, sdlog ≈ 0.46 (~5 day median)

Examples

# COVID-19 default
inc.period.samp(100)
#>   [1]  4.304586  7.562345  2.704031  6.693057  5.299185  6.748075  6.282093
#>   [8]  2.469091  8.171096  4.308251  4.914612  5.207512  5.431148  3.098084
#>  [15]  9.456431  6.050932  4.029526  7.274511  2.376243  5.747199  2.647439
#>  [22]  7.415069  2.336703  5.288458  3.707036  3.344784  7.153647  9.085574
#>  [29]  2.196242  3.249797  9.863261  8.847146  9.317403  2.495189 10.190495
#>  [36]  5.111858  4.908933  6.364363  5.443713  3.369933  3.967767  2.809986
#>  [43]  3.504817 10.568953  3.372648  5.899616  4.014743  3.771924 10.591535
#>  [50]  5.500504  2.492633  5.077649  4.590009  3.244074  1.784117 13.153399
#>  [57]  3.145387  4.848641  5.754486  5.261675  1.718088  4.811794  5.398642
#>  [64]  5.124040 13.051094 15.019934  7.278008  7.489422  4.374946  8.465516
#>  [71]  3.223521  6.764511  5.996892  6.130873  8.979261  3.187545  5.691187
#>  [78] 10.357451  4.212825  4.678403  4.568239  3.080943  6.490650 11.383941
#>  [85]  2.392856  2.514284  7.912064  6.973144 14.674901  4.271384  2.997468
#>  [92]  8.745666  9.214587  5.636121  4.177853  6.945383 13.695652 13.096319
#>  [99]  2.306404  3.896343

# Influenza (shorter incubation)
inc.period.samp(100, meanlog = 0.34, sdlog = 0.42)
#>   [1] 0.8596983 1.7768187 0.6543415 1.6573244 1.4300313 2.3058721 1.2612372
#>   [8] 0.9015307 1.5272770 1.8463747 1.6232716 2.6131688 1.4480977 3.4318748
#>  [15] 0.7557817 1.6674794 2.5230555 1.6344672 1.8494922 2.1856325 0.6117613
#>  [22] 2.3338563 1.3085274 1.5902838 2.3913619 0.8721061 0.8736787 1.4713748
#>  [29] 2.0101720 1.2732569 3.3750468 1.3409977 0.7827079 0.8695651 2.8746636
#>  [36] 1.3584836 1.1692259 1.3363335 1.9548518 1.1016874 1.3217025 1.7526210
#>  [43] 2.9098537 2.5816689 2.6581785 1.1956823 3.0264568 1.1144639 0.9766602
#>  [50] 1.2160292 2.1954963 1.9768100 2.9965541 1.3442969 2.1225759 0.6841438
#>  [57] 0.9905963 2.2304572 1.3060346 1.5145238 1.0477653 0.9385786 0.9326944
#>  [64] 1.2187160 2.2795676 1.6655285 1.1528217 1.3284789 2.3518602 1.6158102
#>  [71] 1.2143669 1.3479821 1.4256128 1.6523084 1.4172936 1.4094315 2.0770622
#>  [78] 1.0538025 1.6188380 1.1816387 2.0800453 3.0432713 1.0449570 1.4099783
#>  [85] 3.3014434 0.7997084 2.2859261 1.2900129 1.1986885 2.9129509 0.9851079
#>  [92] 0.9381363 2.1540424 0.7495048 0.8541710 1.8307001 3.3957548 1.0857061
#>  [99] 0.7071248 1.3988000