Sample Time to Isolation

Draws n samples from the active detection distribution (contact tracing) for the specified TTIQ scenario or custom configuration.

Usage

iso.time.samp(n.ind, the.scenario = "optimal", scenario_config = NULL)

Arguments

n.ind

Integer. Number of samples to draw.

the.scenario

Character. Built-in scenario name: "optimal", "partial", or "baseline". Ignored if scenario_config is provided.

scenario_config

A scenario_config object created by create_scenario_config. If provided, the.scenario is ignored.

Value

Numeric vector of isolation times in days.

Details

Time to isolation represents the delay from index case infection to when close contacts are identified and isolated via contact tracing (active detection).

Examples

# Using built-in scenario
iso.time.samp(100, the.scenario = "optimal")
#>   [1]  1.6805915  2.5958539  2.4317124  1.7287600  1.6347119  1.6072077
#>   [7]  3.9350375  3.0640952  1.1764712  3.3440614  1.0238240  1.1514189
#>  [13]  4.2909599  2.0600668  4.6398420  2.8025219  4.2908996  2.2547336
#>  [19]  1.4768567  1.3003947  4.2903243  1.8225511  5.7099234  7.8164744
#>  [25]  3.1335669  3.9935501  3.7697977  5.7293089  1.6461304  3.5281641
#>  [31]  4.4838398  1.6067288  1.7828391  2.9161159  0.8397628  3.5529256
#>  [37]  1.8423083  2.5719709  4.0385637  3.8008530  3.2693182  6.2843340
#>  [43]  2.3078697  3.7528533  2.1906504  3.6702881  3.8378521  3.5492059
#>  [49]  2.3193575  5.4650226  5.0698113  3.0738070  2.7456308  3.4511924
#>  [55] 13.4319635  4.0173063  4.7366685  3.6823210  4.1145737  2.8972202
#>  [61]  2.0575796  3.7636192  1.3100252  2.9551743  1.9423699  6.7732149
#>  [67]  2.7049144  3.3089096  6.3552935  2.9686898  2.6405853  2.1319794
#>  [73]  4.1797462  5.6440526  3.0129836  3.4696945  2.0787324  4.2381738
#>  [79]  2.6183743  1.3003947  1.0711717  1.8275694  4.6594852  3.9324429
#>  [85]  5.5845146  1.5258317  2.8335360 11.5844639  2.4717447  2.7201692
#>  [91]  2.9653297  4.8611623  2.4985632  3.2807047  2.8291386  2.7440809
#>  [97]  5.1182252  4.9874708  2.6518103  2.5984422

# Using custom configuration
config <- create_scenario_config(
  active_detection = list(type = "lognormal", meanlog = 1.0, sdlog = 0.4)
)
iso.time.samp(100, scenario_config = config)
#>   [1] 3.054732 2.770752 3.246273 3.550518 2.398908 7.459622 2.298176 4.640758
#>   [9] 2.486086 3.446121 2.648234 2.053459 2.199726 2.909292 2.845232 1.976906
#>  [17] 2.473366 4.368128 2.930527 4.118644 2.270688 2.047345 2.751568 2.270614
#>  [25] 3.554482 5.516192 3.503636 3.149955 4.721183 1.921603 7.203147 2.766373
#>  [33] 2.393346 7.968216 3.414991 2.186074 2.475654 1.512334 3.179772 3.905444
#>  [41] 4.395820 4.358111 1.773802 3.094208 2.516904 3.648391 4.523894 2.454548
#>  [49] 2.707377 1.656248 1.776501 1.791012 1.660546 2.979275 3.485909 2.846088
#>  [57] 4.955568 2.160567 1.699082 2.777626 4.419898 2.278119 3.663407 2.959946
#>  [65] 1.599265 5.299180 2.584443 1.184412 5.307344 2.729240 4.800719 3.025866
#>  [73] 3.194912 1.938958 3.211394 1.513253 2.242869 3.891271 2.208817 1.894370
#>  [81] 4.150287 5.157656 3.991528 2.568690 1.832746 3.049460 2.446497 4.035461
#>  [89] 3.058919 5.049336 2.328568 2.798432 2.332011 2.648396 3.799414 4.645587
#>  [97] 2.177399 4.756041 3.247183 2.908479